requirements.txt
# requirements.txt

discord.py
asyncpg
python-dotenv
